requirements.txt
SpeechRecognition
pyttsx3
pywhatkit
requests
gTTS
groq
pygame
openwakeword
numpy
pyaudio
python-dotenv